+++ application/controllers/Halaman.php
@@ -107,6 +107,56 @@ class Halaman extends CI_Controller {
 		echo $html;
 	}
 
+	public function upload_image() {
+		$imageThumb = $this->input->post('imageThumb');
+	    $search = [
+	    	'data:image/png;base64,',
+	    	'data:image/jpg;base64,',
+	    	'data:image/jpeg;base64,',
+	    	'data:image/gif;base64,'
+	    ];
+	    $find = [
+	    	'',
+	    	'',
+	    	'',
+	    	''
+	    ];
+	    $imageThumb = str_replace($search, $find, $imageThumb);
+	    $imageThumb = str_replace(' ', '+', $imageThumb);
+	    $data = base64_decode($imageThumb);
+	    $newName = random_string(10) . '.jpg';
+	    $patchOriginal = 'uploads/images/page/original/' . $newName;
+	    $patchThumbnail = 'uploads/images/page/thumbnail/' . $newName;
+	    $upload = uploadOriginalImage($data)
+		   			->save($patchOriginal);
+		if (!$upload) {
+			$result = [
+		    	'status' => false,
+		    	'cls' => 'failed',
+				'msg' => 'Upload thumbnail gagal'
+		    ];
+		}
+
+		$uploadThumbnail = uploadThumbImage($data)
+							->save($patchThumbnail);
+		if (!$uploadThumbnail) {
+			$result = [
+		    	'status' => false,
+		    	'cls' => 'failed',
+				'msg' => 'Upload thumbnail gagal'
+		    ];
+		}
+
+		if ($upload && $uploadThumbnail) {
+			$result = [
+		    	'status' => true,
+		    	'value'	 => base_url($patchOriginal),
+		    	'fileName'	 => $newName
+		    ];
+		}
+	    echo json_encode($result);
+	}
+
 	public function cari($id)
 	{
 		$data = $this->halaman_model->get_by_id($id);