+++ application/views/dashboard/_parts/header.php
@@ -7,7 +7,7 @@ scratch. This page gets rid of all links and provides the needed markup only.
     <head>
         <meta charset="utf-8">
         <meta http-equiv="X-UA-Compatible" content="IE=edge">
-        <title>Pelayanan Surat | Desa Tunggul Payung</title>
+        <title><?php echo ambil_pengaturan('nama_website'); ?></title>
         <meta content="width=device-width, initial-scale=1, maximum-scale=1, user-scalable=no" name="viewport">
         <!-- Bootstrap -->
         <link rel="stylesheet" href="<?php echo base_url(); ?>assets/bower_components/bootstrap/dist/css/bootstrap.min.css">
@@ -23,6 +23,8 @@ scratch. This page gets rid of all links and provides the needed markup only.
         <link rel="stylesheet" href="<?php echo base_url(); ?>assets/bower_components/Ionicons/css/ionicons.min.css">
         <!-- Admin LTE CSS -->
         <link rel="stylesheet" href="<?php echo base_url(); ?>assets/dist/css/AdminLTE.min.css">
+        <!-- Admin LTE CSS -->
+        <link rel="stylesheet" href="<?php echo base_url(); ?>assets/dist/css/custom.css">
         <!-- Skin blue -->
         <link rel="stylesheet" href="<?php echo base_url(); ?>assets/dist/css/skins/skin-blue.min.css">
         <link rel="stylesheet"